+++ src/main/kotlin/com/github/insanusmokrassar/SauceNaoAPI/Launcher.kt
@@ -2,15 +2,13 @@ package com.github.insanusmokrassar.SauceNaoAPI
 
 import kotlinx.coroutines.*
 
-fun main(vararg args: String) {
+suspend fun main(vararg args: String) {
     val (key, requestUrl) = args
 
-    runBlocking {
-        val api = SauceNaoAPI(key, scope = GlobalScope)
-        api.use {
-            println(
-                it.request(requestUrl)
-            )
-        }
+    val api = SauceNaoAPI(key, scope = GlobalScope)
+    api.use {
+        println(
+            it.request(requestUrl)
+        )
     }
 }